convert f32 to f64

+++ native/src/consts.rs
@@ -1,12 +1,12 @@
 pub const RIGHT_SIDE_WIDTH: f32 = 200.;
 pub const LEFT_SIDE_WIDTH: f32 = 200.;
-pub const CANVAS_INIT_SCALE: f32 = 2.;
+pub const CANVAS_INIT_SCALE: f64 = 2.;
 pub const TOOLBOX_H: f32 = 220.;
 
 pub const COLOR_PICKER_H: f32 = 200.;
 
-pub const DEFAULT_WIDTH: f32 = 480.;
-pub const DEFAULT_HEIGHT: f32 = 270.;
+pub const DEFAULT_WIDTH: f64 = 480.;
+pub const DEFAULT_HEIGHT: f64 = 270.;
 
 // pub const DEFAULT_WIDTH: f32 = 100.;
 // pub const DEFAULT_HEIGHT: f32 = 100.;
