+++ xprite/src/algorithms/polyline.rs
@@ -21,7 +21,7 @@ impl Polyline {
     }
 
     /// line simplification algorithm
-    pub fn reumann_witkam(&self, tol: f32) -> Result<Polyline, String> {
+    pub fn reumann_witkam(&self, tol: f64) -> Result<Polyline, String> {
         if self.pos.len() < 3 {
             return Err("polyline has fewer than 10 points".to_owned());
         }
@@ -79,7 +79,7 @@ impl Polyline {
 }
 
 /// distance from p0 to p1--p2
-pub fn point_line_distance(p0: Vec2f, p1: Vec2f, p2: Vec2f) -> f32 {
+pub fn point_line_distance(p0: Vec2f, p1: Vec2f, p2: Vec2f) -> f64 {
     ((p2.x - p1.x) * (p1.y - p0.y) - (p1.x - p0.x) * (p2.y - p1.y)).abs()
-        / ((p2.x - p1.x) * (p2.x - p1.x) + (p2.y - p1.y) * (p2.y as f32 - p1.y as f32)).sqrt()
+        / ((p2.x - p1.x) * (p2.x - p1.x) + (p2.y - p1.y) * (p2.y as f64 - p1.y as f64)).sqrt()
 }
