+++ xprite/src/rendering/traits.rs
@@ -3,20 +3,20 @@ pub enum MouseCursorType {
 }
 
 pub trait Renderer {
-    fn width(&self) -> f32;
-    fn height(&self) -> f32;
-    fn rect(&mut self, p0: [f32; 2], p1: [f32; 2], color: [f32; 4], filled: bool);
-    fn pixel(&mut self, x: f32, y: f32, color: [f32; 4], filled: bool);
-    fn circ(&mut self, p0: [f32; 2], r: f32, color: [f32; 4], filled: bool);
-    fn line(&mut self, p0: [f32; 2], p1: [f32; 2], color: [f32; 4]);
+    fn width(&self) -> f64;
+    fn height(&self) -> f64;
+    fn rect(&mut self, p0: [f64; 2], p1: [f64; 2], color: [f32; 4], filled: bool);
+    fn pixel(&mut self, x: f64, y: f64, color: [f32; 4], filled: bool);
+    fn circ(&mut self, p0: [f64; 2], r: f64, color: [f32; 4], filled: bool);
+    fn line(&mut self, p0: [f64; 2], p1: [f64; 2], color: [f32; 4]);
     fn bezier(
         &mut self,
-        p0: [f32; 2],
-        cp1: [f32; 2],
-        cp2: [f32; 2],
-        p1: [f32; 2],
+        p0: [f64; 2],
+        cp1: [f64; 2],
+        cp2: [f64; 2],
+        p1: [f64; 2],
         color: [f32; 4],
-        thickness: f32,
+        thickness: f64,
     );
     #[allow(unused)]
     fn set_mouse_cursor(&mut self, cursor_type: MouseCursorType) {}
