+++ xprite/src/tools/rect.rs
@@ -34,7 +34,7 @@ impl Rect {
                 let y1 = stop.point.y;
                 let dx = x1 - x0;
                 let dy = y1 - y0;
-                let d = f32::min(dx, dy);
+                let d = f64::min(dx, dy);
                 let mut end = start;
                 end.point.x = start.point.x + d;
                 end.point.y = start.point.y + d;
